Isolate and move common portable ELF loading from fixtures into //flutter/testing.
Also update all known test harnesses to use this and fixes the broken shell_unittests harness.

Fixes https://github.com/flutter/flutter/issues/49853
